Time: 3 ms (46.4%), Space: 8.7 MB (42.35%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         if (x < 0) return false;  
 
         int original = x;
-        long long rev = 0; 
+         long rev = 0; 
 
         while (x != 0) {
             int ld = x % 10;
